feat(review-loop): status line with round, activity, issues, and tokens
