+++ review-loop/src/live-renderer.ts
@@ -3,15 +3,9 @@
 // Use of this software is governed by the Business Source License 1.1.
 // See LICENSE in the project root for details.
 
-import path from 'node:path'
-
 import { formatDecidedLine, formatFoundLine } from './issue-format.js'
+import { activitySummary, formatDuration, formatTokenCount, MIDDLE_DOT, truncate } from './live-format.js'
 import type { IssueProgressEvent, ProgressReporter, UsageDelta } from './progress-log.js'
-
-const ELLIPSIS = '\u2026'
-const ARROW = '\u25B6'
-const CHECK = '\u2713'
-const MIDDLE_DOT = '\u00B7'
 
 const ERASE_LINE = '\u001b[2K'
 const CURSOR_DOWN = '\u001b[1B'
@@ -24,94 +18,6 @@ export interface RendererStream {
   write(chunk: string): boolean
   isTTY?: boolean
   columns?: number
-}
-
-export function formatDuration(ms: number): string {
-  const totalSeconds = Math.max(0, Math.floor(ms / 1000))
-  if (totalSeconds < 60) {
-    return `${totalSeconds}s`
-  }
-  const minutes = Math.floor(totalSeconds / 60)
-  const seconds = totalSeconds % 60
-  return `${minutes}m${seconds.toString().padStart(2, '0')}s`
-}
-
-function truncate(value: string, max: number): string {
-  if (max <= 0) {
-    return ''
-  }
-  if (value.length <= max) {
-    return value
-  }
-  return `${value.slice(0, max - 1)}${ELLIPSIS}`
-}
-
-function pickString(obj: Record<string, unknown>, keys: readonly string[]): string {
-  for (const key of keys) {
-    const value = obj[key]
-    if (typeof value === 'string' && value.length > 0) {
-      return value
-    }
-  }
-  return ''
-}
-
-function firstStringValue(obj: Record<string, unknown>): string {
-  for (const value of Object.values(obj)) {
-    if (typeof value === 'string' && value.length > 0) {
-      return value
-    }
-  }
-  return ''
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === 'object'
-}
-
-export function formatToolArg(tool: string, input: unknown): string {
-  const obj = isRecord(input) ? input : {}
-  switch (tool) {
-    case 'read':
-    case 'edit':
-    case 'write': {
-      const filePath = pickString(obj, ['filePath', 'path'])
-      return filePath === '' ? '' : path.basename(filePath)
-    }
-    case 'bash':
-      return truncate(pickString(obj, ['command']), 40)
-    case 'grep':
-    case 'glob':
-      return truncate(pickString(obj, ['pattern']), 40)
-    case 'task':
-      return truncate(pickString(obj, ['description', 'subagent_type']), 40)
-    default:
-      return truncate(firstStringValue(obj), 40)
-  }
-}
-
-export function formatLiveLine(
-  label: string,
-  tool: string,
-  arg: string,
-  elapsedMs: number,
-  toolCount: number,
-  status = '',
-): string {
-  const toolPart = tool === '' ? 'thinking' : arg === '' ? tool : `${tool} ${arg}`
-  const tools = `${toolCount} tool${toolCount === 1 ? '' : 's'}`
-  const suffix = status === '' ? '' : ` ${MIDDLE_DOT} ${status}`
-  return `  ${label.padEnd(10)} ${ARROW} ${toolPart} ${MIDDLE_DOT} ${formatDuration(elapsedMs)} ${MIDDLE_DOT} ${tools}${suffix}`
-}
-
-export function formatStepFooter(
-  label: string,
-  elapsedMs: number,
-  toolCount: number,
-  tokens: { input: number; output: number },
-): string {
-  const tools = `${toolCount} tool${toolCount === 1 ? '' : 's'}`
-  return `  ${label} ${CHECK} ${formatDuration(elapsedMs)} ${MIDDLE_DOT} ${tools} ${MIDDLE_DOT} in ${tokens.input} / out ${tokens.output}`
 }
 
 export async function withLivePhase<T>(
@@ -247,8 +153,22 @@ export class LiveRenderer implements ProgressReporter {
 
   private statusLine(): string {
     if (this.slots.size === 0) return ''
-    const suffix = this.statusSuffix()
-    return suffix === '' ? '' : `  ${'status'.padEnd(10)} ${suffix}`
+    const parts: string[] = []
+    if (this.round > 0) parts.push(`round ${this.round}/${this.maxRounds}`)
+    const activity = activitySummary(this.slots.keys())
+    if (activity !== '') parts.push(activity)
+    if (this.startedAt !== 0) parts.push(formatDuration(Date.now() - this.startedAt))
+    const segments: string[] = []
+    if (this.counts.open > 0) segments.push(`${this.counts.open} open`)
+    if (this.counts.fixed > 0) segments.push(`${this.counts.fixed} fixed`)
+    if (this.counts.rejected > 0) segments.push(`${this.counts.rejected} rejected`)
+    if (this.counts.needsHuman > 0) segments.push(`${this.counts.needsHuman} needs human`)
+    if (segments.length > 0) parts.push(`issues: ${segments.join(` ${MIDDLE_DOT} `)}`)
+    if (this.usageTotals.input > 0 || this.usageTotals.output > 0) {
+      parts.push(`in ${formatTokenCount(this.usageTotals.input)} / out ${formatTokenCount(this.usageTotals.output)}`)
+    }
+    if (parts.length === 0) return ''
+    return `  ${'status'.padEnd(10)} ${parts.join(` ${MIDDLE_DOT} `)}`
   }
 
   private renderBlock(): void {
