+++ tests/review-loop/live-renderer.test.ts
@@ -5,76 +5,7 @@
 
 import { describe, expect, test } from 'bun:test'
 
-import {
-  formatDuration,
-  formatLiveLine,
-  formatStepFooter,
-  formatToolArg,
-  LiveRenderer,
-  type RendererStream,
-} from '../../review-loop/src/live-renderer.js'
-
-describe('formatDuration', () => {
-  test('formats seconds under a minute', () => {
-    expect(formatDuration(0)).toBe('0s')
-    expect(formatDuration(42000)).toBe('42s')
-  })
-  test('formats minutes and seconds', () => {
-    expect(formatDuration(125000)).toBe('2m05s')
-  })
-})
-
-describe('formatToolArg', () => {
-  test('read/edit/write use basename of filePath', () => {
-    expect(formatToolArg('read', { filePath: '/a/b/cli.ts' })).toBe('cli.ts')
-    expect(formatToolArg('edit', { path: '/a/b/src/x.ts' })).toBe('x.ts')
-  })
-  test('bash truncates command', () => {
-    expect(formatToolArg('bash', { command: 'echo hi' })).toBe('echo hi')
-    const long = 'x'.repeat(60)
-    expect(formatToolArg('bash', { command: long })).toHaveLength(40)
-  })
-  test('grep/glob use pattern', () => {
-    expect(formatToolArg('grep', { pattern: 'TODO' })).toBe('TODO')
-  })
-  test('task uses description then subagent_type', () => {
-    expect(formatToolArg('task', { description: 'find files' })).toBe('find files')
-    expect(formatToolArg('task', { subagent_type: 'explore' })).toBe('explore')
-  })
-  test('fallback uses first string value', () => {
-    expect(formatToolArg('custom', { a: 'hello', b: 'world' })).toBe('hello')
-  })
-  test('empty input yields empty string', () => {
-    expect(formatToolArg('read', {})).toBe('')
-    expect(formatToolArg('mystery', {})).toBe('')
-  })
-})
-
-describe('formatLiveLine', () => {
-  test('renders label, tool, arg, elapsed, count', () => {
-    const line = formatLiveLine('fixer', 'edit', 'cli.ts', 42000, 3)
-    expect(line).toContain('fixer')
-    expect(line).toContain('edit cli.ts')
-    expect(line).toContain('42s')
-    expect(line).toContain('3 tools')
-  })
-  test('singular tool count', () => {
-    expect(formatLiveLine('reviewer', 'read', 'a.ts', 1000, 1)).toContain('1 tool')
-  })
-  test('no tool yet shows thinking', () => {
-    expect(formatLiveLine('reviewer', '', '', 2000, 0)).toContain('thinking')
-  })
-})
-
-describe('formatStepFooter', () => {
-  test('renders summary with tokens', () => {
-    const footer = formatStepFooter('reviewer', 18000, 4, { input: 13373, output: 31 })
-    expect(footer).toContain('reviewer')
-    expect(footer).toContain('18s')
-    expect(footer).toContain('4 tools')
-    expect(footer).toContain('in 13373 / out 31')
-  })
-})
+import { LiveRenderer, type RendererStream } from '../../review-loop/src/live-renderer.js'
 
 function makeStream(opts: { isTTY?: boolean; columns?: number } = {}): {
   stream: RendererStream
@@ -163,18 +94,6 @@ describe('LiveRenderer', () => {
     r.live(['line 1', 'line 2'])
     expect(captured.join('')).toContain('line 1')
     expect(captured.join('')).toContain('line 2')
-  })
-})
-
-describe('formatLiveLine status suffix', () => {
-  test('appends status after tool count when provided', () => {
-    const line = formatLiveLine('fixer', 'read', 'cli.ts', 5000, 3, 'round 1/3 · issues: 2 open')
-    expect(line).toContain('3 tools · round 1/3 · issues: 2 open')
-  })
-  test('omits suffix segment when status is empty', () => {
-    const line = formatLiveLine('fixer', 'read', 'cli.ts', 5000, 3)
-    expect(line).toContain('3 tools')
-    expect(line).not.toContain('round')
   })
 })
 
@@ -275,7 +194,9 @@ describe('LiveRenderer slots', () => {
     r.slot('b', 'line-b')
     r.slot('c', 'line-c')
     const redraw = output[2]!
-    expect(redraw.startsWith('\r\u001b[1A')).toBe(true)
+    // statusLine renders one block line above the slots, so the block is now 4 lines
+    // ([status, line-a, line-b, line-c]) and the third redraw moves the cursor up 2.
+    expect(redraw.startsWith('\r\u001b[2A')).toBe(true)
     expect(redraw).toContain('line-a')
     expect(redraw).toContain('line-c')
   })
@@ -315,5 +236,49 @@ describe('LiveRenderer slots', () => {
     expect(() => r.event('x')).not.toThrow()
     expect(r.dynamic).toBe(false)
     expect(() => r.slot('a', 'line')).not.toThrow()
+  })
+})
+
+describe('status line', () => {
+  test('combines round, activity, issues, and tokens', () => {
+    const { stream, output } = makeStream({ isTTY: true, columns: 200 })
+    const r = new LiveRenderer(stream)
+    r.issue({ type: 'round', round: 1, maxRounds: 2 })
+    r.issue({
+      type: 'found',
+      id: 'aaaaaaaa-0000-0000-0000-000000000000',
+      severity: 'low',
+      file: 'a.ts',
+      line: 1,
+      title: 'A',
+    })
+    r.usage({ input: 228819, output: 9824, reasoning: 49844, cost: 0 })
+    r.slot('fixer-w1', 'x')
+    r.slot('fixer-w2-retry', 'y')
+    const status = output[output.length - 1]!.split('\n')[0]!
+    expect(status).toContain('status')
+    expect(status).toContain('round 1/2')
+    expect(status).toContain('fix×2')
+    expect(status).toContain('issues: 1 open')
+    expect(status).toContain('in 228.8k / out 9.8k')
+  })
+
+  test('maps slot keys to activity verbs', () => {
+    const { stream, output } = makeStream({ isTTY: true, columns: 200 })
+    const r = new LiveRenderer(stream)
+    r.slot('reviewer', 'x')
+    const status = output[output.length - 1]!.split('\n')[0]!
+    expect(status).toContain('review')
+    expect(status).not.toContain('reviewer')
+  })
+
+  test('accumulates usage across calls', () => {
+    const { stream, output } = makeStream({ isTTY: true, columns: 200 })
+    const r = new LiveRenderer(stream)
+    r.usage({ input: 500, output: 0, reasoning: 0, cost: 0 })
+    r.usage({ input: 600, output: 100, reasoning: 0, cost: 0 })
+    r.slot('a', 'x')
+    const status = output[output.length - 1]!.split('\n')[0]!
+    expect(status).toContain('in 1.1k / out 100')
   })
 })
